darwin

+++ .zshrc
@@ -24,20 +24,24 @@ if [[ ! -f $ZSH/plugins/omz/sudo.plugin.zsh ]]; then
   curl https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/plugins/sudo/sudo.plugin.zsh -o  $ZSH/plugins/omz/sudo.plugin.zsh
 fi
 
-if [[ -z "$(fc-list 'JetBrainsMono Nerd Font')" ]]; then
-  mkdir -p ~/.local/
-  mkdir -p ~/.local/share/
-  mkdir -p ~/.local/share/fonts/
-  mkdir -p ~/.local/share/fonts/jetbrains/
-  cd ~/.local/share/fonts/jetbrains/
-  curl -OL 'https://github.com/ryanoasis/nerd-fonts/releases/download/v3.0.2/JetBrainsMono.tar.xz'
-  xz -d JetBrainsMono.tar.xz
-  tar -xf JetBrainsMono.tar
-  rm JetBrainsMono.tar
-  cd 
+
+if [[ $(uname) != "Darwin" ]]; then
+  if [[ -z "$(fc-list 'JetBrainsMono Nerd Font')" ]]; then
+    mkdir -p ~/.local/
+    mkdir -p ~/.local/share/
+    mkdir -p ~/.local/share/fonts/
+    mkdir -p ~/.local/share/fonts/jetbrains/
+    cd ~/.local/share/fonts/jetbrains/
+    curl -OL 'https://github.com/ryanoasis/nerd-fonts/releases/download/v3.0.2/JetBrainsMono.tar.xz'
+    xz -d JetBrainsMono.tar.xz
+    tar -xf JetBrainsMono.tar
+    rm JetBrainsMono.tar
+    tar -xf JetBrainsMono.tar
+    rm JetBrainsMono.tar
+   cd 
+  fi
 fi
 
-  
 
 source $ZSH/themes/powerlevel10k/powerlevel10k.zsh-theme
 source $ZSH/plugins/zsh-autocomplete/zsh-autocomplete.plugin.zsh
@@ -64,13 +68,23 @@ zstyle ':autocomplete:history-search-backward:*' list-lines 256
 [[ ! -f ~/.p10k.zsh ]] || source ~/.p10k.zsh
 
 export EDITOR=vim
-export ANDROID_HOME=$HOME/Android/Sdk
-export ANDROID_SDK_ROOT=$HOME/Android/Sdk
-export ANDROID_AVD_HOME=$HOME/.android/avd
-export PATH=$PATH:/usr/local/bin:$HOME/.cargo/bin:$ANDROID_SDK_ROOT/platform-tools:$ANDROID_SDK_ROOT/tools:/usr/local/go/bin:$HOME/.local/bin
-export HISTFILE=~/.zsh_history
-export HISTSIZE=10000000
-export SAVEHIST=$HISTSIZE
+
+if [[ $(uname) == "Darwin" ]]; then
+  export ANDROID_HOME=$HOME/Library/Android/Sdk
+  export ANDROID_SDK_ROOT=$HOME/Library/Android/Sdk
+else 
+  export ANDROID_SDK_ROOT=$HOME/Android/Sdk
+  export ANDROID_HOME=$HOME/Android/Sdk
+fi
+
+
+
+
+if [[ $(uname) == "Darwin" ]]; then
+  export PATH="/opt/homebrew/opt/openjdk@11/bin:$PATH"
+  eval "$(/opt/homebrew/bin/brew shellenv)"
+fi
+
 
 # bun completions
 [ -s "$HOME/.bun/_bun" ] && source "$HOME/.bun/_bun"
@@ -78,6 +92,16 @@ export SAVEHIST=$HISTSIZE
 # bun
 export BUN_INSTALL="$HOME/.bun"
 export PATH="$BUN_INSTALL/bin:$PATH"
+
+# Turso
+export PATH="$HOME/.turso:$PATH"
+
+export ANDROID_AVD_HOME=$HOME/.android/avd
+export PATH="$PATH:/usr/local/bin:$HOME/.cargo/bin:$ANDROID_SDK_ROOT/platform-tools:$ANDROID_SDK_ROOT/emulator:/$ANDROID_SDK_ROOT/tools:/usr/local/go/bin:$HOME/.local/bin"
+export HISTFILE=~/.zsh_history
+export HISTSIZE=10000000
+export SAVEHIST=$HISTSIZE
+
 
 if (( $+commands[pacman] ))
 then
